+++ packages/expo-widgets/android/src/main/java/expo/modules/widgets/ui/ButtonView.kt
@@ -5,121 +5,156 @@ import androidx.compose.ui.graphics.Color
 import androidx.glance.Button
 import androidx.glance.ButtonDefaults
 import androidx.glance.GlanceTheme
+import androidx.glance.action.Action
+import androidx.glance.action.action
+import androidx.glance.appwidget.components.FilledButton
+import androidx.glance.appwidget.components.OutlineButton
 import androidx.glance.unit.ColorProvider
 import com.facebook.react.bridge.ReadableArray
 import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.ReadableType
-import expo.modules.ui.button.ButtonProps
+import expo.modules.kotlin.views.ComposeProps
+import expo.modules.kotlin.views.OptimizedComposeProps
+import expo.modules.ui.ModifierList
+import expo.modules.ui.button.ButtonColors
 import expo.modules.ui.colorToComposeColorOrNull
-import androidx.glance.ButtonColors
-import androidx.glance.appwidget.components.FilledButton
-import androidx.glance.appwidget.components.OutlineButton
+import expo.modules.widgets.WidgetInteraction
+import expo.modules.widgets.toGlanceAction
+import androidx.glance.ButtonColors as GlanceButtonColors
+
+@OptimizedComposeProps
+data class WidgetButtonProps(
+  val colors: ButtonColors = ButtonColors(),
+  val enabled: Boolean = true,
+  val modifiers: ModifierList = emptyList(),
+  val target: String? = null
+) : ComposeProps
 
 @Composable
 internal fun ButtonView(
-  props: ButtonProps,
-  children: List<ReadableMap> = emptyList()
+  props: WidgetButtonProps,
+  children: List<ReadableMap>,
+  source: String,
 ) {
-  FilledButton(
-    text = props.textContent(children),
-    onClick = {},
-    modifier = props.modifiers.toGlanceModifier(),
-    enabled = props.enabled,
-    colors = props.toGlanceButtonColors(
-      defaultBackgroundColor = GlanceTheme.colors.primary,
-      defaultContentColor = GlanceTheme.colors.onPrimary
-    )
+  WidgetFilledButton(
+    props = props,
+    children = children,
+    onClick = buttonAction(source, props.target),
+    defaultBackgroundColor = GlanceTheme.colors.primary,
+    defaultContentColor = GlanceTheme.colors.onPrimary
   )
 }
 
 @Composable
 internal fun FilledTonalButtonView(
-  props: ButtonProps,
-  children: List<ReadableMap> = emptyList()
+  props: WidgetButtonProps,
+  children: List<ReadableMap>,
+  source: String,
 ) {
-  FilledButton(
-    text = props.textContent(children),
-    onClick = {},
-    modifier = props.modifiers.toGlanceModifier(),
-    enabled = props.enabled,
-    colors = props.toGlanceButtonColors(
-      defaultBackgroundColor = GlanceTheme.colors.secondaryContainer,
-      defaultContentColor = GlanceTheme.colors.onSecondaryContainer
-    )
+  WidgetFilledButton(
+    props = props,
+    children = children,
+    onClick = buttonAction(source, props.target),
+    defaultBackgroundColor = GlanceTheme.colors.secondaryContainer,
+    defaultContentColor = GlanceTheme.colors.onSecondaryContainer
   )
 }
 
 @Composable
 internal fun OutlinedButtonView(
-  props: ButtonProps,
-  children: List<ReadableMap> = emptyList()
+  props: WidgetButtonProps,
+  children: List<ReadableMap>,
+  source: String,
 ) {
+  val buttonText = props.textContent(children)
+  val contentColor = props.toGlanceContentColor(GlanceTheme.colors.primary)
+  val modifier = props.modifiers.toGlanceModifier()
+
   OutlineButton(
-    text = props.textContent(children),
-    contentColor = props.toGlanceContentColor(GlanceTheme.colors.primary),
-    onClick = {},
-    modifier = props.modifiers.toGlanceModifier(),
+    text = buttonText,
+    contentColor = contentColor,
+    onClick = buttonAction(source, props.target),
+    modifier = modifier,
     enabled = props.enabled
   )
 }
 
 @Composable
 internal fun ElevatedButtonView(
-  props: ButtonProps,
-  children: List<ReadableMap> = emptyList()
+  props: WidgetButtonProps,
+  children: List<ReadableMap>,
+  source: String,
 ) {
-  FilledButton(
-    text = props.textContent(children),
-    onClick = {},
-    modifier = props.modifiers.toGlanceModifier(),
-    enabled = props.enabled,
-    colors = props.toGlanceButtonColors(
-      defaultBackgroundColor = GlanceTheme.colors.surface,
-      defaultContentColor = GlanceTheme.colors.primary
-    )
+  WidgetFilledButton(
+    props = props,
+    children = children,
+    onClick = buttonAction(source, props.target),
+    defaultBackgroundColor = GlanceTheme.colors.surface,
+    defaultContentColor = GlanceTheme.colors.primary
   )
 }
 
 @Composable
 internal fun TextButtonView(
-  props: ButtonProps,
-  children: List<ReadableMap> = emptyList()
+  props: WidgetButtonProps,
+  children: List<ReadableMap>,
+  source: String,
 ) {
+  val buttonText = props.textContent(children)
+  val modifier = props.modifiers.toGlanceModifier()
+  val colors = props.toGlanceButtonColors(
+    defaultBackgroundColor = Color.Transparent.toGlanceColorProvider(),
+    defaultContentColor = GlanceTheme.colors.primary
+  )
+
   Button(
+    text = buttonText,
+    onClick = buttonAction(source, props.target),
+    modifier = modifier,
+    enabled = props.enabled,
+    colors = colors
+  )
+}
+
+@Composable
+private fun WidgetFilledButton(
+  props: WidgetButtonProps,
+  children: List<ReadableMap>,
+  onClick: Action,
+  defaultBackgroundColor: ColorProvider,
+  defaultContentColor: ColorProvider
+) {
+  FilledButton(
     text = props.textContent(children),
-    onClick = {},
+    onClick = onClick,
     modifier = props.modifiers.toGlanceModifier(),
     enabled = props.enabled,
-    colors = props.toGlanceButtonColors(
-      defaultBackgroundColor = Color.Transparent.toGlanceColorProvider(),
-      defaultContentColor = GlanceTheme.colors.primary
-    )
+    colors = props.toGlanceButtonColors(defaultBackgroundColor, defaultContentColor)
   )
 }
 
 // TODO(@jakex7): Find a better way to get text
-private fun ButtonProps.textContent(children: List<ReadableMap>): String {
+private fun WidgetButtonProps.textContent(children: List<ReadableMap>): String {
   return children.textContent() ?: ""
 }
 
 private fun List<ReadableMap>.textContent(): String? {
-  return mapNotNull { it.textFromTextNode() }
-    .joinToString(separator = "")
+  return mapNotNull { it.textFromTextNode() }.joinToString(separator = "")
     .takeIf { it.isNotEmpty() }
 }
 
 @Composable
-private fun ButtonProps.toGlanceButtonColors(
+private fun WidgetButtonProps.toGlanceButtonColors(
   defaultBackgroundColor: ColorProvider,
   defaultContentColor: ColorProvider
-): ButtonColors {
+): GlanceButtonColors {
   return ButtonDefaults.buttonColors(
     backgroundColor = toGlanceContainerColor(defaultBackgroundColor),
     contentColor = toGlanceContentColor(defaultContentColor)
   )
 }
 
-private fun ButtonProps.toGlanceContainerColor(defaultColor: ColorProvider): ColorProvider {
+private fun WidgetButtonProps.toGlanceContainerColor(defaultColor: ColorProvider): ColorProvider {
   return colorToComposeColorOrNull(
     if (enabled) {
       colors.containerColor
@@ -129,7 +164,7 @@ private fun ButtonProps.toGlanceContainerColor(defaultColor: ColorProvider): Col
   )?.toGlanceColorProvider() ?: defaultColor
 }
 
-private fun ButtonProps.toGlanceContentColor(defaultColor: ColorProvider): ColorProvider {
+private fun WidgetButtonProps.toGlanceContentColor(defaultColor: ColorProvider): ColorProvider {
   return colorToComposeColorOrNull(
     if (enabled) {
       colors.contentColor
@@ -137,6 +172,11 @@ private fun ButtonProps.toGlanceContentColor(defaultColor: ColorProvider): Color
       colors.disabledContentColor ?: colors.contentColor
     }
   )?.toGlanceColorProvider() ?: defaultColor
+}
+
+@Composable
+private fun buttonAction(source: String, target: String?): Action {
+  return target?.let { WidgetInteraction(source, it).toGlanceAction() } ?: action {}
 }
 
 private fun ReadableMap.textFromTextNode(): String? {